Use uvec4 for mousePicking output
Change fragment shader output from vec4 to uvec4 so the mouse picking pass emits unsigned integer color values. This preserves exact integer components when encoding object IDs via bit shifts and avoids implicit float conversions that could break picking.

+++ app/src/main/cpp/GraphicsEngine/source/shaders/mousePicking.frag
@@ -5,7 +5,7 @@ layout(push_constant) uniform pointPC {
   uint objectID;
 } pc;
 
-layout(location = 0) out vec4 outColor;
+layout(location = 0) out uvec4 outColor;
 
 void main() {
   float r = (pc.objectID >> 16) & 0xFF;
